+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.5] - 2021-07-12
+### Added
+- Menu add sub-item button bugfix
+- Added new `alias_suggester` option to `cms.page_templates` item config. Must be a FQCN of a class that implements `RevisionTen\CMS\Interfaces\AliasSuggesterInterface`. If this option is set, the `alias_prefix` option will be ignored.
+
 ## [3.0.4] - 2021-06-16
 ### Added
 - Added support for `RevisionTen\CMS\Form\Types\DoctrineType` to menu items